Split for URL Focus setters and Focus Selectors into
